Add state db abstraction
