+++ monitoring_service/state_db/__init__.py
@@ -1,5 +1,7 @@
 from .db import StateDB
+from .sqlite_db import StateDBSqlite
 
 __all__ = [
-    'StateDB'
+    'StateDB',
+    'StateDBSqlite'
 ]
