+++ monitoring_service/test/fixtures/server.py
@@ -24,12 +24,12 @@ def blockchain(web3):
 
 
 @pytest.fixture
-def monitoring_service(server_private_key, blockchain, dummy_transport, state_db, web3):
+def monitoring_service(server_private_key, blockchain, dummy_transport, state_db_mock, web3):
     ms = MonitoringService(
         server_private_key,
         transport=dummy_transport,
         blockchain=blockchain,
-        state_db=state_db
+        state_db=state_db_mock
     )
     yield ms
     ms.blockchain.stop()
