+++ monitoring_service/test/mockups/__init__.py
@@ -1,0 +1,5 @@
+from .state_db import StateDBMock
+
+__all__ = [
+    'StateDBMock'
+]
