+++ monitoring_service/test/test_state_db.py
@@ -1,8 +1,8 @@
-def test_state_db(state_db, get_random_bp):
+def test_state_db_sqlite(state_db_sqlite, get_random_bp):
     bp = get_random_bp()
     bp = bp.serialize_data()
-    state_db.store_balance_proof(bp)
-    ret = state_db.balance_proofs
+    state_db_sqlite.store_balance_proof(bp)
+    ret = state_db_sqlite.balance_proofs
     fields_to_check = [
         'contract_address',
         'participant1',
